added printf functionality over sysex, added continue for clock, added tests for printf, added correct song position ptr parsing, fixed callback handling for clock
git-svn-id: https://mididuino.googlecode.com/svn/trunk@1096 6f8c04f6-b8b6-11de-9846-1b058f2a8eb1

+++ hardware/libraries/PageTools/AutoEncoderPage.h
@@ -34,28 +34,28 @@ template <typename EncoderType>
 void AutoEncoderPage<EncoderType>::on32Callback(uint32_t counter) {
   if (muted)
     return;
-  for (int i = 0; i < 4; i++) {
+  for (uint8_t i = 0; i < 4; i++) {
     recEncoders[i].playback(counter & 0xFF);
   }
 }
 
 template <typename EncoderType>
 void AutoEncoderPage<EncoderType>::startRecording() {
-  for (int i = 0; i < 4; i++) {
+  for (uint8_t i = 0; i < 4; i++) {
     recEncoders[i].startRecording();
   }
 }
 
 template <typename EncoderType>
 void AutoEncoderPage<EncoderType>::stopRecording() {
-  for (int i = 0; i < 4; i++) {
+  for (uint8_t i = 0; i < 4; i++) {
     recEncoders[i].stopRecording();
   }
 }
 
 template <typename EncoderType>
 void AutoEncoderPage<EncoderType>::clearRecording() {
-  for (int i = 0; i < 4; i++) {
+  for (uint8_t i = 0; i < 4; i++) {
     recEncoders[i].clearRecording();
   }
 }
@@ -68,7 +68,7 @@ void AutoEncoderPage<EncoderType>::clearRecording(uint8_t i) {
 template <typename EncoderType>
 void AutoEncoderPage<EncoderType>::setup() {
   muted = false;
-  for (int i = 0; i < 4; i++) {
+  for (uint8_t i = 0; i < 4; i++) {
     realEncoders[i].setName("___");
     recEncoders[i].initRecordingEncoder(&realEncoders[i]);
     encoders[i] = &recEncoders[i];
@@ -84,39 +84,39 @@ void AutoEncoderPage<EncoderType>::autoLearnLast4() {
   int8_t encoderAssigned[4] = { -1, -1, -1, -1 };
   incoming_cc_t ccs[4];
 
-  int count = ccHandler.incomingCCs.size();
-  for (int i = 0; i < count; i++) {
+  uint8_t count = ccHandler.incomingCCs.size();
+  for (uint8_t i = 0; i < count; i++) {
     ccHandler.incomingCCs.getCopy(i, &ccs[i]);
     incoming_cc_t *cc = &ccs[i];
-    for (int j = 0; j < 4; j++) {
+    for (uint8_t j = 0; j < 4; j++) {
       if ((realEncoders[j].getCC() == cc->cc) &&
-	  (realEncoders[j].getChannel() == cc->channel)) {
-	ccAssigned[i] = j;
-	encoderAssigned[j] = i;
-	break;
+					(realEncoders[j].getChannel() == cc->channel)) {
+				ccAssigned[i] = j;
+				encoderAssigned[j] = i;
+				break;
       }
     }
   }
 
-  for (int i = 0; i < count; i++) {
+  for (uint8_t i = 0; i < count; i++) {
     incoming_cc_t *cc = &ccs[i];
     if (ccAssigned[i] != -1) {
       if ((realEncoders[ccAssigned[i]].getChannel() != cc->channel) &&
-	  (realEncoders[ccAssigned[i]].getCC() != cc->cc)) {
-	realEncoders[ccAssigned[i]].initCCEncoder(cc->channel, cc->cc);
-	realEncoders[ccAssigned[i]].setValue(cc->value);
-	clearRecording(ccAssigned[i]);
+					(realEncoders[ccAssigned[i]].getCC() != cc->cc)) {
+				realEncoders[ccAssigned[i]].initCCEncoder(cc->channel, cc->cc);
+				realEncoders[ccAssigned[i]].setValue(cc->value);
+				clearRecording(ccAssigned[i]);
       }
     } else {
-      for (int j = 0; j < 4; j++) {
-	if (encoderAssigned[j] == -1) {
-	  ccAssigned[i] = j;
-	  encoderAssigned[j] = i;
-	  realEncoders[ccAssigned[i]].initCCEncoder(cc->channel, cc->cc);
-	  realEncoders[ccAssigned[i]].setValue(cc->value);
-	  clearRecording(ccAssigned[i]);
-	  break;
-	}
+      for (uint8_t j = 0; j < 4; j++) {
+				if (encoderAssigned[j] == -1) {
+					ccAssigned[i] = j;
+					encoderAssigned[j] = i;
+					realEncoders[ccAssigned[i]].initCCEncoder(cc->channel, cc->cc);
+					realEncoders[ccAssigned[i]].setValue(cc->value);
+					clearRecording(ccAssigned[i]);
+					break;
+				}
       }
     }
   }
@@ -140,13 +140,13 @@ bool AutoEncoderPage<EncoderType>::handleEvent(gui_event_t *event) {
     return true;
   }
   if (BUTTON_DOWN(Buttons.BUTTON4)) {
-    for (int i = Buttons.ENCODER1; i <= Buttons.ENCODER4; i++) {
+    for (uint8_t i = Buttons.ENCODER1; i <= Buttons.ENCODER4; i++) {
       if (EVENT_PRESSED(event, i)) {
-	GUI.setLine(GUI.LINE1);
-	GUI.flash_string_fill("CLEAR");
-	GUI.setLine(GUI.LINE2);
-	GUI.flash_put_value(0, i);
-	clearRecording(i);
+				GUI.setLine(GUI.LINE1);
+				GUI.flash_string_fill("CLEAR");
+				GUI.setLine(GUI.LINE2);
+				GUI.flash_put_value(0, i);
+				clearRecording(i);
       }
     }
   }
